backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "userdb FastAPI backend"
readme = "README.md"
authors = [{ name = "Tim Yates", email = "timyates100@gmail.com" }]
requires-python = ">=3.13"

dependencies = [
    "boto3>=1.42.41",
    "botocore[crt]>=1.42.41",
    "fastapi[standard]>=0.116.1",
    "humps>=0.2.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "sqlmodel>=0.0.24",
]

[tool.pytest.ini_options]
pythonpath = ["./src", "."]
asyncio_mode = "auto"
